+++ src/ui/panels/equationManagePanel/box_filters.tsx
@@ -2,6 +2,7 @@ import { EquationMatch } from "@/utils/parsers/equation_parser";
 
 /**
  * Here we use a loose match 
+ * @remarks we may use `\Large\boxed{equation}`, so even not start with 
  * @abstract only check if the first non-blank line of equations is start with `\boxed`.
  * @param equation
  * @param enableTypstMode if true, it will check `#box` instead of `\boxed`
@@ -11,13 +12,15 @@ export function boxedEquationFilter(
     equation: EquationMatch,
     enableTypstMode: boolean,
     skipFirstLine: boolean,
+    typstBoxSymbol = "#box",
 ): boolean {
-    const matchPrefix = enableTypstMode ? "#box" : String.raw`\boxed`;
     const lines = equation.content.split("\n")
         .map(line => line.trim())
         .filter(line => line.length > 0);
     if (lines.length === 0) return false;
     const lineIndex = (skipFirstLine && lines.length > 1) ? 1 : 0;
     const targetLine = lines[lineIndex];
-    return targetLine.startsWith(matchPrefix);
+    return enableTypstMode ? 
+       targetLine.includes(typstBoxSymbol)
+       : targetLine.includes(String.raw`\boxed`);
 }